+++ backend/modules/blog/actions/add.php
@@ -148,7 +148,11 @@ class BackendBlogAdd extends BackendBaseActionAdd
 					// the image path
 					$imagePath = FRONTEND_FILES_PATH . '/blog/images';
 
-					// validate the image
+					// create folders if needed
+					if(!SpoonDirectory::exists($imagePath . '/source')) SpoonDirectory::create($imagePath . '/source');
+					if(!SpoonDirectory::exists($imagePath . '/128x128')) SpoonDirectory::create($imagePath . '/128x128');
+
+					// image provided?
 					if($this->frm->getField('image')->isFilled())
 					{
 						// build the image name